mount: Fix possibility to disable CRC checking
One of the previous commits added checking for CRC validity
when reading data. There is an ENABLE_CRC flag in the build system,
which should be respected by this code, but was not. This commit
fixes this issue: code compiled with ENABLE_CRC=NO does not
verify checksums during read operations.

Change-Id: I9d7fea39a78bf46a37dd118eeff6fec0c1c055ad