tests: Add test_petabyte_file
We have introduced checking of appending data to the 1PetaByte file.
We use sparse file so we do not need to allocate all space to perform this test.
In addition we also make sure if the sparse fille is correct
(returns zeros from not allocated blocks)
The last think we check is if appended data occupies not more storage
than one CHUNK_SIZE.