tests: Add test_petabyte_file
We have introduced checking of appending data to the a 1 petabyte file.
We use sparse file so we do not need to allocate all space to perform this test.
In addition we also make sure if the sparse file is correct (returns zeros from
not allocated blocks). The last think we check is if the appended data does not
occupy more storage than one CHUNK_SIZE.

Closes #109