chunkserver: Fix bug in network thread
Add new status for CSTOCS_GET_CHUNK_BLOCKS request in
chunkserver network thread, because it isn't an 'IDLE' job.
Now other requests received by a chunkserver on the same socket
before the response for CSTOCS_GET_CHUNK_BLOCKS is sent will be
considered as invalid and connection will be closed.

Change-Id: I8f7c9a5797a4d1bb0ec1021c39193770a5b06ea6