mount: Make writing xor chunks fault-tolernat
This commit modifies chunk writer to use it's journal in the way, that
if any exception occures during the write operation, xor parity is always
consistent with all the parts and the parts contain proper data.

Writing many different xor levels of one chunks is not implemented (ie.
it's not possible to write chunk which has both xor2 and xor3 copies,
or standard and xor copies).

Change-Id: Icb8c9a31bb760cb5bd214fbbf89b1658ff5906c0