master: Add support for refreshing locks
In the master server add support for refreshing locks acquired on chunks.
This is implemented by expanding the LIZ_CLTOMA_FUSE_WRITE_CHUNK message by
lockId, so that client will be able to refresh his lock providing an ID of
his previously acquired lock. This mechanism will be useful for writing xor
chunks. In case of some writeoperation fail (because of eg. problems with
a connection to a chunkserver) the parity may be temporarily inconsistent
with the data. We may want not to release the chunk lock until we fix the
inconsistency.

Temporarily all lock ID's are equal to 1. Using different values will be
added in the future.

Change-Id: Ic0bcd2548693763ebbfbeb658612d53a95df34fd