master, metarestore: Add full support for lock IDs
This commits makes the master server assign a random lock ID for every
locked chunk. This ID has to be used when client wants to refresh his lock
or unlock it. It makes lock refreshing safe (with lock IDs it is very
unlikely that any client will successfully acquire a lock on a chunk
locked by other client).

Change-Id: I701e93f3753f2aa48fd2826cee419320c48736e0