mount: Add support for lock ID's in mount
If a recoverable error occurs during writing chunk, it's still
possible that write operation can be completed successfully soon.
This commit modifies write worker in mount not to release chunk locks
in case of recoverable errors and refresh the lock using it's ID
on the next try. It also introduces a functional test of the chunk
locking mechanism.

Change-Id: I6d1bb33e499e4ca595e9e335037891ad2be611dd