master: Fix writing lost chunks
Consider the following scenario: there is only part 1/2 (version 1)
available, two other are on chunkservers which are temporarily down.
When client wants to write this chunk the master server can't just
increase a version of available copies and let the client to write it,
because this would destoy the chunk -- we will have part 1/2 (version 2)
and two other invalid parts (version 1). This commit fixes this issue.
Master server will not increase a version of a chunk if it can't be
written.

Change-Id: I5a6c3c197e08854d5c83b24e6ae0fb79c1f24d7a