tests: Fix descriptor leak test to work with valgrind
The test tried to obtain chunkserver pid using 'pidof mfschunkserver',
but this doesn't work when chunkserver is being run under valgrind.

Moreover, the default timeout multiplier of 5 turned out to be too low.

Use 'mfschunkserver test' instead of pidof and increase the timeout
multiplier to 10 to fix the above issues.