master: Fix versions of temporary throwaway chunks
When chunkserver reports a chunk with ID unknown to master, master
creates a chunk with this ID in its data structures and lets chunk
jobs take care of removing it (as an unused chunk).

These temporary chunks were getting wrong IDs when copies reported
by the chunkservers were marked "to delete". This commit fixes it.