mount: Don't block write workers when data arrives slowly
ChunkWriter used to try to assemble whole stripes and block the write
worker when there is some stripe being assembled (ie. more data blocks
are needed), but nothing is actually being written to chunkservers.
This commit modifies write worker to stop processing a chunk if nothing
can be immediately sent.

Change-Id: Ib3bbb4e509cc4e16bec9dbdcf5310ef27de8ba87