mount: Add a possibility to use I/O limiting without cgroups
This commit introduces support for a special entry in the I/O limit config
file, ie:

  limit unclassified <some_limit>

If this is the only entry in the config file, it will be used as a limit for
all processes. If there is also a subsystem and some limits for cgroups
specicifed, it will be used as a limit for processes from cgroups which are
not present in the configuration.