master, chunkserver, metalogger: Add isalive command
This commit introduces a new command isalive which can be use to test
if a daemon (mfsmaster, mfsmetalogger of mfschunkserser) if running.
The exit status of this command is as follows:
* 0 if the daemon is running
* 1 if the daemon is not running
* other if an error occured during the check

The exit status 1 is reserved to the isalive command and is used only
to indicate that the daemon is down. No other invocation (including
wrong usage, eg. mfsmaster isaliv) will exit with the status of 1.