mount: Disable support for capability-based security
Kernel on some Linux distributions issues getxattr("security.capability")
request every 64 kB of written data. This is a big problem for the master
server if we want to write a lot of data. This commit fixes this issue in
the most simple way by returning ENOTSUP when trying to use capabilities on
files. In the future we may consider implementing some cache for these calls
or a configuration option which will allow users to decide what to do with
capabilities.

Change-Id: I850d1cd59147eb8b2d9510e0e4df6f16282c0bb8