tests: Add tests of manipulating permissions and ACLs
Add tests which verify correctness of operations which manipulate permissions
when ACLs are used on LizardFS.

The first test (test_acl_group_class) is very simple and verifies if setfacl
and getfacl work properly with LizardFS, ie:
* if getfacl returns what it should return
* if modifying permissions using chmod influences output of getfacl
* if modifying permissions using setfacl influences file's access mask
'default:' entries are not covered by this test.

The test_acl_behavior is a little more complicated: it tries to do the same
things in two directory trees: one on LizardFS and the other one on
a filesystem which natively supports ACLs (the one in in $TEMP_DIR). After
each step (setting some permissions using chmod or setfacl, copying files,
etc.) it verifies if access masks and results of getfacl are the same.

Change-Id: I991a3dae64679400b20c82bc03a3f08150be4d86