utils: Fix dissection of ACL messages in the Wireskark plugin
The right name for a 16-bit field in LIZ_CLTOMA_SETACL and LIZ_MATOCL_GETACL
which represents an access mask is 'mode'. This name is used in
the AccessControlList class and many other messages, ie. CLTOMA_FUSE_MKDIR
or CLTOMA_FUSE_MKNOD.

Moreover, fields named 'mode' are displayed in octal in the Wireshark
interface, which is the expected behavior in case of these two messages.

Change-Id: I939200aef96c80b146f5715a07aad60b9330e128