common: Add ACL xattr <-> AccessControlList converter
ACL on Linux is stored as serialized struct of kernel in extended attributes,
so it's very inconvenient to use them directly. To allow an easy manipulation
on these data, this commit introduces a class which converts them to an
AccessControlList object (and vice versa). A definition of kernel structs have
been extracted from its sources and adapted to this use in this case.

Change-Id: I7e218ad349db410d0ae1ca42fa7b91bc95fd11f1