metarestore: Fix restore may fail due bug in auto mode
The restore may fail if the metadata file from master is 
older than from metalogger.

Details

The metadata lastversion is read from wrong offset 8 while
it is stored as bigendian 64 bit integer at offset 12.

The lastversion loaded is used for selecting best metadata file when
performing 'metarestore -a'. The high DWORD of lastversion loaded due
the bug is actually lastinode stored as 32-bit integer under offset 8.
The low DWORD of lastversion loaded is actually the high DWORD of
lastversion stored.

Change-Id: I48ee57c054a5f7b097512fa30a2ad3d943046620