master, mount: Fix checking permissions during 'touch'
man utime says:

  The utime() system call changes the access and modification times of the
  inode specified by filename to the actime and modtime fields of times
  respectively.
  If times is NULL, then the access and modification times of the file are
  set to the current time.
  Changing timestamps is permitted when: either the process has appropriate
  privileges, or the effective user ID equals the user ID of the file, or
  times is NULL and the process has write permission for the file.

The master server couldn't distinguish between the NULL and non-NULL
version, so it was always checking for the ownership of the changed
file. This commit reuses two no longer used setattr flags to pass the
required information to the master server and modifies the master server
to check permissions correctly.

Change-Id: I63ff90e6714a7383b016f2d54b0ecf848bc15424