master: Fix checking for exceeding hard quota limits
Hard quota limit is considered exceeded if the actual usage is greater
than or equal to the hard limit, while soft quota limit is considered
exceeded if the usage is strictly greater than the soft limit. The
implementation was always checking for the 'greater than' condition and
this commit changes it to reflect the proper behavior.

Change-Id: Ifffb2372703f4aca4d8166e36a04383a63647d92