tests: Fix failing test_probe_list_chunkservers
This test was failing very frequently when using valgrind because of
a race: it was waiting until all the chunkservers report at least one
chunk to the master server and was checking if they reported 7 chunk
total. It often happened that some chunkserver reports its first chunk
before the second one and in a meanwhile the test finished waiting and
crashed.

Change-Id: I405652ca8832b86ff3a7dd94ac692baed8070783