chunkserver: Fix sending status of a failed replication
Due to the bug in definition of the LIZ_CSTOMA_REPLICATE message
function overwriteStatusField used by the chunkserver was damaging the
network packet (it was overwriting part of the chunk's version) after
the replication was finished. The bug didn't appear when replication of
chunk with version less than 65536 succeeded, so it was quite rare.

This commit fixes the bug by changing order of fields in the packet's
definition (in MFSCommunication.h the order was already right), so that
overwriteStatusField will work with this message. It also changes the
name of the message to be consistent with other CS<->MA messages (eg.
createChunk, deleteChunk, duplicateChunk).

Code after this change will not be backward compatible with the code
from the previous commit but this is legit, because there are no
existing installations running any code which is compatible with the
previous commit.

Change-Id: I5ffcc598de24cf15e3ab1e150786562a8df279ec