master: Fix removing spare chunk copies
It turned out that when cleaning up redundant copies of single chunk,
the master server was omitting some copies if there were more than one stored
on single chunkserver. It caused the master server to be loaded unnecessarily
while chunk copies were still remaining on chunkserver. This commit fixes it.

Change-Id: Ic972d574bfaed69ba9ece77c7b2b7485ce0501cf