mount: Retry after truncate failure
Unlike in, for instance, reading, truncating a file immediately failed in case
of its non-availability. This commit makes mount retry to truncate a file.

The mount will try every min(60 seconds, 2 ** (n - 1) * 200 miliseconds) where
`n' is the number of the try (which is not greater than the `mfsioretries'
parameter passed to mount).

Change-Id: Id92b1342f5457aace65aa1bb06e3d0c747613c62