mount: Add subclass of WriteChunkLocator for truncates
This commit adds a class which is suppose to set the existing lock on chunk
before locking it during locating. It'll be needed when xor chunks are
truncated down. In this case the master will lock a chunk and delegate a client
to write zeros to a specific part of chunk.

Change-Id: If2f19c715efc611cb5dbda0cfb67b56018c8d6c5