mount: Fix bug in writing xor chunks with missing parts
A bug in ChunkWriter caused mfsmount to abort if there was a need to
read a block of data (to fill stripe before writing it) and this block
had to be recovered from parity in a stripe that is shorter than xor
level (eg. last stripe of xor3 or xor7 chunk). In such a case
ChunkWriter was trying to read blocks 1024, 1025, and so on which makes
some assertions to fail.

Change-Id: I4e8d94c04f586099f79dd2537f73f6f84f503284