chunkserver: Fix creating new chunks by duplicating existing
This commit fixes a bug in duplicate and duptrunc functions in
chunkserver. New chunks that were created by these functions had wrong
signatures, ie. instead of:

  LIZC 1.0 <id> <vesion> <type>

chunkserver was creating chunk files with the following signature:

  MFSC 1.0 <id> <version>

This commit unifies code that writes the signature and uses it in
create, duplicate, and duptrunc operations.

Change-Id: I5c2546d7355f6fdb27eaa90072daf48691073eff