mount: Fix bug when handling getxattr(file, name, NULL, 0)
When calling getxattr("filename", "attrname", NULL, 0) we should return
length of the extended attribute without returning its value. This
behavior was broken in mfsmount and this commit fixes it and extends
test_xattr to verify if this feature works properly. To do this,
a simple C++ interpreter is added to the testing framework.

Change-Id: I9352dcb3ad7c45bb4cae90a2b85bd3b0e5ee60c6