master: Add AUTO_RECOVERY option
This commit introduces a possibility to apply changelogs in the master
server's process. This feature makes it possible to start master server
without using mfsmetarestore after a crash, ie. when changes from
changelog.0.mfs need to be applied to the metadata.mfs file. When there
is a AUTO_RECOVERY = 1 option in the master server's configuration file
set, it will automatically perform this task before starting if it's
needed (ie. if there is a stale metadata.mfs.lock file left). If
AUTO_RECOVERY = 0, it will refuse to start in such a case.

Change-Id: Ib9509c5dba130c52c3c7d15605a41c74a865e3d0