master: Fix handling of removed IO limiting groups
There is a short race window when mount can queue an IO limiting request
for group which is being removed from the master configuration at the
exact same time.

When this happened, mount was repeatedly sending this packet and getting
disconnected by master, until the transmission retry counter has reached
its limit.

Make master respond with 0 to requests for groups which don't exist to
avoid the repeated retransmissions and disconnections.

Change-Id: Idca1b3992d454ed844bd715b1ce4024e8aff1722