tests: Separate "private" and "public" parts of lizardfs.sh
This commit makes it clear which functions and global variables from
lizardfs.sh should be called and which should not by appending "_" to
all the "private" members. It also removes usage of such members from
other tests, especially from test_moosefs_upgrade which was failing
because of using such a wrong function.

Change-Id: I599ba9e6d888e24ad91ac5cd4a0472af0caf7748