mount: Fix listing directories with a lot of entries
Due tu a buggy handling of the 'size' parameter LizardClient::readdir
was returning the whole listing of a directory each time fuse asked for
a small piece of it (eg. 4 KiB). This made the complexity of the
getdents syscall O(n^2). This commit fixes the issue and renames some
variables to better describe their roles in the code. It also adds
a test of listing directories with a lot of entries.

Change-Id: Ibf1ef2afa6f56f9f5cdaa8df66438eb14a4afcce