master: Add destructors for fsedge, fsnode, xattr_data_entry
This commit adds destructors for structs which are used to represent
metadata and need some memory deallocation. Constructors are also
provided to initialize fields that destructor takes care of. All the
allocation of these structs are changed from malloc/free to new/delete.
This commit prepares the code to be able to easily deallocate all the
metadata.

Change-Id: I5b9642abf70f2913930d4160ba19de237f022244