utils: Fix handling LIZ_MATOCL_METADATASERVER_STATUS in wireshark
This message was using 'status' field which caused it to be rendered in
wireshark as:

  LIZ_MATOCL_METADATASERVER_STATUS msgid:1 status:ERROR_ENOTDIR metadata_version:2

This issue is fixes by creating a new field 'serverstatus' with
a proper dictionary.

Also creating a new field 'metadata_version' is not necessary, because an
existing field 'metadataversion' can be used in this message.

Change-Id: I9193cd1f3f7ff42a16bb9e0ab2fffea71de83bee