master: Add MAGIC_AUTO_FILE_REPAIR option
This option may be useful in tests which fail because of killing the
master server just after it decided to increase version of some chunk.
This may result losing all copies of such chunk and only mfsfilerepair
would be able to recover it. When MAGIC_AUTO_FILE_REPAIR is set, master
will call fs_repair before writing a file if its chunk is lost. This
feature is enabled in test_changing_master_during_build.

Change-Id: I2d411ff3c6ec8f15e03db8ea842c9075a6d10d6e