cmake: Handle Boost libraries properly
Since Boost is always required to build LizardFS, there is no need to
check for it every time we want to use it. But there is a need to add
its location to to the list of directories to be searched for header
files, because the compilation fails if boost is found in a non-standard
path.

Change-Id: Ia3e39a6d752610ffa0acc3bfae83c7d66542c4e4