common, mount: Introduce ReadPlanner for multi-variant plans
This commit introduces MultiVariantReadPlanner, which is used by
mfsmount to generate plans when reading chunks. This class is the only
planner which uses scores and it uses them to choose which variant of
the plan should be the 'basic' one. In this commit the planner doesn't
actually generate anything more than the basic variant, but this commit
modifies interface of the ReadPlanner class not to use any scores and
modifies the code which uses it to deal with this change. It also
implements as much as needed to provide full backwards compatibility
with the previous version of code (which passes scores to
StandardChunkReadPlanner) to make all the functional tests pass.

Change-Id: I229576ad0db83877a65c32a88e2311ee1ce6f154