mount: Add possibility to use additional variants of read plans
This commit adds options which can be used to influence when mountpoint
starts to execute additional read operations when reading data in order
to make the whole operation faster. It also modifies ChunkReader to
instruct planner how to plan future operations on a chunk if there were
timeouts encountered during reading from the chunk before.

It also removes (at least temporarily) a possibility to handle in
parallel two parallel read requests to the same inode.

Change-Id: Ib90db5ec919c4844abc56ca937a0874389791c13