common: Fix bugs in ReadPlanExecutor
The executor was hanging when both basicTimeout and totalTimeout expired
in the same moment, because it couldn't start additional read
operations:

  if (!additionalOperationsStarted && basicTimeout.expired() && !totalTimeout.expired()) {
    (...)
    additionalOperationsStarted = true;

nor break the whole process:

  int status = poll(pollFds.data(), pollFds.size(), pollTimeout);
  (...)
  } else if (status == 0 && additionalOperationsStarted) {
    // The time is out, our chunkservers appear to be completely unresponsive.

The second bug is that the executor was checking for the

  plan_->isReadingFinished(unfinishedOperations)

condition before filling unfinishedOperations set which caused the
executor to finish too early.

The third bug is that exceptions thrown when a connection with
a chunkserver was lost during a read operation were not handled
properly.

The fourth bug is that descriptorsForBasicReadOperations variable was
updated for all operations, not only for basic operations.

The fifth bug is that file descriptors where leaking in case of some
errors.

Change-Id: I6c7d012dfc0fd7793723ecfbf8c99460da02a513