master: Add possibility to recalculate FS checksum in background
The process of recalculating checksum of the filesystem metadata can be
run in background to avoid stopping the master server for a couple of
seconds if anything went wrong. It tries to fix errors calculating the
checksum once again. During the recalculation CHECKSUM entries are
not added to changelog.

Also config entry 'METADATA_CHECKSUM_RECALCULATION_SPEED' is added to
adjust background recalculation speed.

Change-Id: I06e94ca8eb6488cf5e4a663c3e466f332292b24c