common: Change format of debug logs
If LIZARDFS_LOG_ORIGIN feature is enabled the format of debug logs
changes from:

  tag: some messgae

to:

  tag: path/file.cc:812:some_function: some message

Some tests try to parse debug logs for specific messages and fail when
LIZARDFS_LOG_ORIGIN is enabled. This commit modifies debug logs to look
like this when origins are logged:

  # path/file.cc:812:some_function
  tag: some message

This shouldn't break any tests regardless if origins are enabled or not.

Change-Id: If5c079208314ad42630bd38b698bd8d09e2eddca