chunkserver: Change chunk format breaking backward compatibility
This commit changes the chunk format. The previous format of chunk was

[<Header>:1KB] [(<Block checksum>:4B) * 1024] [(<Block data>:64KB) * N]

Where N was the number of currently created blocks. After the change the
format is

[(<Block checksum><Block data>:(4B + 64KB)) * N]

New chunks end with '.liz' extension instead of previously used '.mfs'.

The new format reduces the number of disk seeks performed during
sequential file write. It also reduces the number of IO operations
needed to do a single random read. As for other operations it's almost
the same.

Sparse chunks are still supported by interpreting empty block and empty
CRC as valid.

Backward compability will be restored with a further commit.

Change-Id: I5abf15c0420c695a5f541a15ce6d12b2b4e41f9c