master: Avoid losing chunks during emergency incversion
When some chunk operation is interrupted by a disconnection of
a chunkserver, the master server increases version of the chunk on all
other chunkservers. This works with standard copies, but in case of xor
copies we can do this only if we have enough parts left to be able to
read and write such a chunk in the future. This commit changes
a condition under which we can perform the emergency version increase
from valid_copies>0 to !c->isLost() to deal with the problem.

Change-Id: Icdb0b7449ecbac23b98a6f237dd2ce8f20430273