mount: Don't keep partial stripes in write buffers forever
This commit modifies behavior of mfsmount which was trying to optimize
out partial stripe writes so aggressively that it could keep some data
in its buffer forever. The previous algorithm was to write a partial
stripe only if:
* flush is requested by calling fsync, closing a descriptor, reading
  data from the same descriptor or calling truncate on the file
* the same descriptor if written at an offset which doesn't belong to
  the last written block.
This could lead to keeping data in buffers forever if a process opens
a descriptor, writes some data and doesn't do anything more with it.
This commit modifies mfsmount to write such a partial stripe after
a couple of seconds of inactivity.

Change-Id: Id7882885d410fde59f268f61591a7c7513593d70