utils: Add generating dictionaries by make_dissector.py
This commit modifies the Wireshark plugin generator to generate
dictionaries for fields by parsing the MFSCommunication.h file. Now the
only thing that needs to be done to list possible values for a given
field is to add something like this in the MFSCommunication.h file:

 /// field values: somefielfd
 #define VALUE_FIRST  2
 #define VALUE_SECOND 4
 #define VALUE_THIRD  10

The list has to end with a blank line or a line which contains only a //
comment. This will instruct Wireshark how to parse expressions like:

 lizardfs.somefield == VALUE_SECOND

and how to print values of lizardfs.somefield in the Info column and in
the protocol tree.

Change-Id: I63f0fd839da660455df3bf8d5d6b4ff93370d7e3