cgi: Fix table headers in the Exports and Mounts tabs
These two tables were missing one column in the header row which made
rest of the rows not consistent with the headers and thus difficult to
read (one of the sesflags was displayed in the "map root/uid" column,
"map root/gid" was displayed in the "map users/uid" column, etc.).

Also modify the sources of the cgi script to use 2.5.0 instead of 1.7.0
as the first version of the master server which support quotas.

Change-Id: Ie906e37d81171768b14eb33f9dc3a2975bcc0e60